+++ lib/read/cardano-wallet-read.cabal
@@ -49,6 +49,7 @@ library
     Cardano.Read.Ledger.Block.BlockNo
     Cardano.Read.Ledger.Block.HeaderHash
     Cardano.Read.Ledger.Block.SlotNo
+    Cardano.Read.Ledger.Block.Txs
     Cardano.Read.Ledger.Tx.Inputs
     Cardano.Read.Ledger.Tx.Outputs
     Cardano.Read.Ledger.Tx.TxId
@@ -78,7 +79,6 @@ library
     Cardano.Wallet.Read.Block.Gen.Build
     Cardano.Wallet.Read.Block.Gen.Byron
     Cardano.Wallet.Read.Block.Gen.Shelley
-    Cardano.Wallet.Read.Block.Txs
     Cardano.Wallet.Read.Chain
     Cardano.Wallet.Read.Eras
     Cardano.Wallet.Read.Eras.EraFun
